+++ clientele/settings.py
@@ -1,6 +1,6 @@
 import platform
 
-VERSION = "1.8.1"
+VERSION = "1.9.0"
 
 
 def split_ver():