+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change log
 
-## UNRELEASED 1.9.0
+## 1.9.0
 
 ### Retry support